emotion-recognition-app/backend/requirements.txt
# Core
fastapi==0.95.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI/ML
torch==2.0.0
torchvision==0.15.1
opencv-python-headless==4.7.0.72
numpy==1.24.2

# Note: dlib should be installed via conda:
# conda install -c conda-forge dlib

# Data Processing
pandas==2.0.0
python-dateutil==2.8.2

# Web & API
websockets==11.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pydantic==1.10.7
python-multipart==0.0.6
python-slugify==8.0.1

# Development
typing-extensions==4.5.0
mypy==1.3.0
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.0
